Removed unnecessary conversion

+++ hdf5_wrapper/stats.py
@@ -71,7 +71,7 @@ def bit_aliasing(reads: list[Read]) -> list[float]:
     reads_values = [read.bits_flatted for read in reads]
     # TODO numpy only
     reads_length = len(reads)
-    return [float(num/reads_length) for num in np.sum(reads_values, axis=0)]
+    return [num/reads_length for num in np.sum(reads_values, axis=0)]
 
 def stable_bit(reads: list[Read]) -> list[float]:
     pass
